refactor(test): move `TransactionEvent`-specic serialization tests to separate class
