+++ src/test/java/it/pagopa/ecommerce/commons/documents/v1/serialization/TransactionEventTypeResolverTest.java
@@ -1,0 +1,100 @@
+package it.pagopa.ecommerce.commons.documents.v1.serialization;
+
+import com.azure.core.util.serializer.JsonSerializer;
+import com.azure.core.util.serializer.TypeReference;
+import io.vavr.control.Either;
+import it.pagopa.ecommerce.commons.documents.v1.*;
+import it.pagopa.ecommerce.commons.generated.server.model.TransactionStatusDto;
+import it.pagopa.ecommerce.commons.queues.QueueEvent;
+import it.pagopa.ecommerce.commons.queues.StrictJsonSerializerProvider;
+import it.pagopa.ecommerce.commons.v1.TransactionTestUtils;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+import reactor.core.publisher.Mono;
+import reactor.test.StepVerifier;
+
+import static it.pagopa.ecommerce.commons.queues.TracingInfoTest.MOCK_TRACING_INFO;
+
+@ExtendWith(MockitoExtension.class)
+public class TransactionEventTypeResolverTest {
+    private final JsonSerializer jsonSerializer = new StrictJsonSerializerProvider().createInstance();
+
+    @Test
+    public void canRoundtripQueueEventSerialization() {
+        QueueEvent<TransactionRefundRequestedEvent> originalEvent = new QueueEvent<>(
+                new TransactionRefundRequestedEvent(
+                        TransactionTestUtils.TRANSACTION_ID,
+                        new TransactionRefundedData(TransactionStatusDto.REFUND_REQUESTED)
+                ),
+                MOCK_TRACING_INFO
+        );
+        byte[] serialized = jsonSerializer.serializeToBytes(originalEvent);
+
+        Mono<QueueEvent<TransactionRefundRequestedEvent>> refundRequestDeserialized = jsonSerializer
+                .deserializeFromBytesAsync(
+                        serialized,
+                        new TypeReference<>() {
+                        }
+                );
+        Mono<QueueEvent<TransactionActivatedEvent>> activatedEventDeserialized = jsonSerializer
+                .deserializeFromBytesAsync(
+                        serialized,
+                        new TypeReference<>() {
+                        }
+                );
+
+        Mono<Either<QueueEvent<TransactionRefundRequestedEvent>, QueueEvent<TransactionActivatedEvent>>> roundTripWithFailure = activatedEventDeserialized
+                .map(Either::<QueueEvent<TransactionRefundRequestedEvent>, QueueEvent<TransactionActivatedEvent>>right)
+                .onErrorResume(
+                        (e) -> refundRequestDeserialized
+                                .map(Either::left)
+                );
+
+        StepVerifier.create(roundTripWithFailure)
+                .expectNextMatches(e -> e.fold(ev -> {
+                    TransactionRefundRequestedEvent event = ev.event();
+                    TransactionRefundRequestedEvent originalTransactionEvent = originalEvent.event();
+                    return event.getData().equals(originalTransactionEvent.getData());
+                }, r -> {
+                    throw new RuntimeException("Event deserialized as incorrect type!");
+                }))
+                .verifyComplete();
+    }
+
+    @Test
+    public void serializerDisambiguatesTypesWithSameShape() {
+        QueueEvent<TransactionClosedEvent> originalEvent = new QueueEvent<>(
+                TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.KO),
+                MOCK_TRACING_INFO
+        );
+        byte[] serialized = jsonSerializer.serializeToBytes(originalEvent);
+
+        String serializedString = new String(serialized);
+        System.out.println("Serialized object: " + serializedString);
+
+        Mono<QueueEvent<TransactionClosedEvent>> closedDeserialized = jsonSerializer
+                .deserializeFromBytesAsync(
+                        serialized,
+                        new TypeReference<>() {
+                        }
+                );
+        Mono<QueueEvent<TransactionClosureFailedEvent>> closureFailedDeserialized = jsonSerializer
+                .deserializeFromBytesAsync(
+                        serialized,
+                        new TypeReference<>() {
+                        }
+                );
+
+        Mono<Either<QueueEvent<TransactionClosedEvent>, QueueEvent<TransactionClosureFailedEvent>>> roundTripWithFailure = closureFailedDeserialized
+                .map(Either::<QueueEvent<TransactionClosedEvent>, QueueEvent<TransactionClosureFailedEvent>>right)
+                .onErrorResume(
+                        (e) -> closedDeserialized
+                                .map(Either::left)
+                );
+
+        StepVerifier.create(roundTripWithFailure)
+                .expectNextMatches(v -> v.getLeft().event().getData().equals(originalEvent.event().getData()))
+                .verifyComplete();
+    }
+}
