+++ src/test/java/it/pagopa/ecommerce/commons/queues/StrictJsonSerializerProviderTest.java
@@ -151,82 +151,4 @@ public class StrictJsonSerializerProviderTest {
 
         assertEquals(UnrecognizedPropertyException.class, exception.getCause().getClass());
     }
-
-    @Test
-    public void canRoundtripQueueEventSerialization() {
-        QueueEvent<TransactionRefundRequestedEvent> originalEvent = new QueueEvent<>(
-                new TransactionRefundRequestedEvent(
-                        TransactionTestUtils.TRANSACTION_ID,
-                        new TransactionRefundedData(TransactionStatusDto.REFUND_REQUESTED)
-                ),
-                MOCK_TRACING_INFO
-        );
-        byte[] serialized = jsonSerializer.serializeToBytes(originalEvent);
-
-        Mono<QueueEvent<TransactionRefundRequestedEvent>> refundRequestDeserialized = jsonSerializer
-                .deserializeFromBytesAsync(
-                        serialized,
-                        new TypeReference<>() {
-                        }
-                );
-        Mono<QueueEvent<TransactionActivatedEvent>> activatedEventDeserialized = jsonSerializer
-                .deserializeFromBytesAsync(
-                        serialized,
-                        new TypeReference<>() {
-                        }
-                );
-
-        Mono<Either<QueueEvent<TransactionRefundRequestedEvent>, QueueEvent<TransactionActivatedEvent>>> roundTripWithFailure = activatedEventDeserialized
-                .map(Either::<QueueEvent<TransactionRefundRequestedEvent>, QueueEvent<TransactionActivatedEvent>>right)
-                .onErrorResume(
-                        (e) -> refundRequestDeserialized
-                                .map(Either::left)
-                );
-
-        StepVerifier.create(roundTripWithFailure)
-                .expectNextMatches(e -> e.fold(ev -> {
-                    TransactionRefundRequestedEvent event = ev.event();
-                    TransactionRefundRequestedEvent originalTransactionEvent = originalEvent.event();
-                    return event.getData().equals(originalTransactionEvent.getData());
-                }, r -> {
-                    throw new RuntimeException("Event deserialized as incorrect type!");
-                }))
-                .verifyComplete();
-    }
-
-    @Test
-    public void serializerDisambiguatesTypesWithSameShape() {
-        QueueEvent<TransactionClosedEvent> originalEvent = new QueueEvent<>(
-                TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.KO),
-                MOCK_TRACING_INFO
-        );
-        byte[] serialized = jsonSerializer.serializeToBytes(originalEvent);
-
-        String serializedString = new String(serialized);
-        System.out.println("Serialized object: " + serializedString);
-
-        Mono<QueueEvent<TransactionClosedEvent>> closedDeserialized = jsonSerializer
-                .deserializeFromBytesAsync(
-                        serialized,
-                        new TypeReference<>() {
-                        }
-                );
-        Mono<QueueEvent<TransactionClosureFailedEvent>> closureFailedDeserialized = jsonSerializer
-                .deserializeFromBytesAsync(
-                        serialized,
-                        new TypeReference<>() {
-                        }
-                );
-
-        Mono<Either<QueueEvent<TransactionClosedEvent>, QueueEvent<TransactionClosureFailedEvent>>> roundTripWithFailure = closureFailedDeserialized
-                .map(Either::<QueueEvent<TransactionClosedEvent>, QueueEvent<TransactionClosureFailedEvent>>right)
-                .onErrorResume(
-                        (e) -> closedDeserialized
-                                .map(Either::left)
-                );
-
-        StepVerifier.create(roundTripWithFailure)
-                .expectNextMatches(v -> v.getLeft().event().getData().equals(originalEvent.event().getData()))
-                .verifyComplete();
-    }
 }
